+++ CHANGES.rst
@@ -7,7 +7,8 @@ Here you can see the full list of changes between each SQLAlchemy-JSON-API relea
 0.4.0 (2016-xx-xx)
 ^^^^^^^^^^^^^^^^^^
 
-- Added type formatting
+- Added type formatting (#5)
+- Added limit and offset parameters (#6)
 
 
 0.3.0 (2015-08-18)